Fixing missing ObjectIdentifier receiving IHaveRequest
If a device was not already cached and a IHaveRequest was received the Objectidentifier was not present, because the Remoteobject including his Objectidentifier was never created and stored to the cache. To prevent this situation now we simple set the ObjectName property and the RemoteObject including his Objectidentifier will be crated and stored automatic.